server/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   strawberry-graphql-django
django==5.2
    # via
    #   ci-and-cd-damo (pyproject.toml)
    #   strawberry-graphql-django
graphql-core==3.2.6
    # via strawberry-graphql
gunicorn==23.0.0
    # via ci-and-cd-damo (pyproject.toml)
packaging==25.0
    # via
    #   gunicorn
    #   strawberry-graphql
python-dateutil==2.9.0.post0
    # via strawberry-graphql
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
strawberry-graphql==0.266.1
    # via strawberry-graphql-django
strawberry-graphql-django==0.59.1
    # via ci-and-cd-damo (pyproject.toml)
typing-extensions==4.13.2
    # via strawberry-graphql
